requirements.txt
beautifulsoup4==4.11.1
Django==3.2.13
djangorestframework==3.13.1
psycopg2==2.9.3
Celery==5.2.6
django-celery-beat==2.2.1
django-createsuperuserwithpassword==2.0.0
requests==2.27.1
lxml==4.9.0
python-dateutil==2.8.2
